requirements.txt
torch
torchvision
tensorflow
keras
timm
efficientnet_pytorch
scikit-learn
opencv-python
matplotlib
seaborn
pandas
numpy
tqdm
Pillow
albumentations
fastapi
uvicorn[standard]
python-multipart